+++ sdk/internal/test/tracingvalidator/validator.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package tracing
+package tracingvalidator
 
 import (
 	"context"
